Adding Problem 242. Valid Anagram
Problem Description: An Anagram is a word or phrase formed by rearranging the letters of a different word or phrase, typically using all the original letters exactly once.

Currently this version of the program does not work for all test cases. Particularly complaining on "aacc"
"ccac". Will investigate, fix and submit.